server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
google-generativeai==0.8.3
python-multipart==0.0.12
email-validator==2.2.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
